+++ hr_applicant/__manifest__.py
@@ -2,7 +2,7 @@
 
 {
     "name": "Applicant Process",
-    "version": "14.0.0.1.0",
+    "version": "15.0.0.1.0",
     "category": "Human Resources",
     "sequence": 90,
     "license": "AGPL-3",
