+++ hr_applicant/models/__init__.py
@@ -1,5 +1,10 @@
 # See LICENSE file for full copyright and licensing details.
-
 from . import hr_recruitment
+from . import hr_app_med_occ
+from . import hr_app_rel_edu
+from . import hr_app_trav_lang
 from . import hr_recruitment_employee
+from . import hr_emp_med_occ
+from . import hr_emp_rel_edu
+from . import hr_emp_trav_lang
 from . import training
