+++ hr_applicant/report/report_view.xml
@@ -1,14 +1,6 @@
-<?xml version="1.0" encoding="utf-8"?>
-
+<?xml version="1.0" encoding="utf-8" ?>
 <odoo>
-
-    <report id="applicant_profile_form_report"
-           model="hr.applicant"
-           string="Resume"
-           name="hr_applicant.applicant_profile"
-           report_type="qweb-pdf"/>
-
-    <record id="applicant_profile_report_paperformat" model="report.paperformat">
+    <record id="paperformat_applicant_profile_report" model="report.paperformat">
         <field name="name">Applicant Profile format</field>
         <field name="default" eval="True" />
         <field name="format">A4</field>
@@ -22,8 +14,15 @@
         <field name="header_line" eval="False" />
         <field name="header_spacing">5</field>
         <field name="dpi">90</field>
-        <field name="report_ids" eval="[(4, ref('hr_applicant.applicant_profile_form_report'))]">
-        </field>
     </record>
-
+    <record id="action_applicant_profile_form_report" model="ir.actions.report">
+        <field name="name">Applicant</field>
+        <field name="model">hr.applicant</field>
+        <field name="report_type">qweb-pdf</field>
+        <field name="report_name">hr_applicant.applicant_profile</field>
+        <field name="report_file">hr_applicant.applicant_profile</field>
+        <field name="paperformat_id" ref="paperformat_applicant_profile_report" />
+        <field name="binding_model_id" ref="model_hr_applicant" />
+        <field name="binding_type">report</field>
+    </record>
 </odoo>
