+++ hr_applicant/static/description/index.html
@@ -30,7 +30,7 @@
         <div class="oe_span12 " style="margin-top:30px">
             <img class="img-border oe_screenshot" src="Record.png" style="width:730px">
         </div>
-        <div class="oe_span12" style="margin-top:40px;">
+       <!--  <div class="oe_span12" style="margin-top:40px;">
             <center>
                 <h4 class="oe_slogan" style="font-size:28px;">
                     <b>
@@ -41,8 +41,8 @@
                     </b>
                 </h4>
             </center>
-        </div>
-        <div>
+        </div> -->
+      <!--   <div>
             <table>
                 <tr>
                     <td class="text-top">
@@ -58,7 +58,7 @@
         </div>
         <div class="oe_span12 " style="margin-top:30px">
             <img class="img-border oe_screenshot" src="attachment.png" style="width:730px">
-        </div>
+        </div> -->
         <div class="oe_span12" style="margin-top:40px;">
             <center>
                 <h4 class="oe_slogan" style="font-size:28px;">
