+++ hr_applicant/wizard/select_training_view.xml
@@ -1,22 +1,36 @@
-<?xml version="1.0" encoding="utf-8"?>
+<?xml version="1.0" encoding="utf-8" ?>
 <odoo>
     <record model="ir.ui.view" id="select_training_wizard_form_view">
         <field name="name">select.training.wizard.form.view</field>
         <field name="model">select.training</field>
         <field name="arch" type="xml">
-            <form string="Need Training" version='9.0'>
+            <form string="Need Training">
                 <group>
-                    <field name="is_triaing_needed"/>
-                    <field name="training_courses_ids" attrs="{'invisible': [('is_triaing_needed', '!=', True)]}" widget="many2many_checkboxes"/>
+                    <field name="is_triaing_needed" />
+                    <field
+                        name="training_courses_ids"
+                        attrs="{'invisible': [('is_triaing_needed', '!=', True)]}"
+                        widget="many2many_checkboxes"
+                    />
                 </group>
                 <footer>
-                    <button icon="fa-check" name="action_done" string="Ok" type="object" class="oe_highlight"/>
-                    <button icon="fa-times" string="Cancel" class="oe_link" special="cancel"/>
+                    <button
+                        icon="fa-check"
+                        name="action_done"
+                        string="Ok"
+                        type="object"
+                        class="oe_highlight"
+                    />
+                    <button
+                        icon="fa-times"
+                        string="Cancel"
+                        class="oe_link"
+                        special="cancel"
+                    />
                 </footer>
             </form>
         </field>
     </record>
-
     <record id="action_select_training_wizard" model="ir.actions.act_window">
         <field name="name">Need Training</field>
         <field name="res_model">select.training</field>
